pyproject.toml
[project]
name = "aqi-bridge"
version = "1.0.0"
description = "BLE-to-WebSocket Bridge for AQI Drone"
requires-python = ">=3.12"
dependencies = [
    "bleak==0.21.1",
    "fastapi==0.110.0",
    "uvicorn[standard]==0.29.0",
    "pydantic==2.6.3",
    "websockets==12.0",
    "python-json-logger==2.0.7",
    "pytest-asyncio==0.23.5",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["apps/aqi-bridge/tests"]
filterwarnings = [
    "ignore:The 'app' shortcut is now deprecated:DeprecationWarning",
]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = ["E501"]  # line length handled by formatter
